# Add doc-values aggregator for the geometric mean aggregation.