--- a/server/src/main/java/io/crate/execution/engine/aggregation/impl/GeometricMeanAggregation.java
+++ b/server/src/main/java/io/crate/execution/engine/aggregation/impl/GeometricMeanAggregation.java
@@ -27,21 +27,35 @@ import io.crate.breaker.RamAccounting;
 import io.crate.common.collections.Lists2;
 import io.crate.data.Input;
 import io.crate.execution.engine.aggregation.AggregationFunction;
+import io.crate.execution.engine.aggregation.DocValueAggregator;
 import io.crate.memory.MemoryManager;
 import io.crate.metadata.functions.Signature;
+import io.crate.types.ByteType;
 import io.crate.types.DataType;
 import io.crate.types.DataTypes;
+import io.crate.types.DoubleType;
 import io.crate.types.FixedWidthType;
+import io.crate.types.FloatType;
+import io.crate.types.IntegerType;
+import io.crate.types.LongType;
+import io.crate.types.ShortType;
+import io.crate.types.TimestampType;
 import org.apache.commons.math3.util.FastMath;
+import org.apache.lucene.index.DocValues;
+import org.apache.lucene.index.LeafReader;
+import org.apache.lucene.index.SortedNumericDocValues;
+import org.apache.lucene.util.NumericUtils;
 import org.elasticsearch.Version;
 import org.elasticsearch.common.breaker.CircuitBreakingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.io.stream.Writeable;
+import org.elasticsearch.index.mapper.MappedFieldType;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
 
 public class GeometricMeanAggregation extends AggregationFunction<GeometricMeanAggregation.GeometricMeanState, Double> {
 
@@ -51,7 +65,7 @@ public class GeometricMeanAggregation extends AggregationFunction<GeometricMeanA
         DataTypes.register(GeometricMeanStateType.ID, in -> GeometricMeanStateType.INSTANCE);
     }
 
-    private static final List<DataType<?>> SUPPORTED_TYPES = Lists2.concat(
+    static final List<DataType<?>> SUPPORTED_TYPES = Lists2.concat(
         DataTypes.NUMERIC_PRIMITIVE_TYPES, DataTypes.TIMESTAMPZ);
 
     public static void register(AggregationImplModule mod) {
@@ -123,6 +137,23 @@ public class GeometricMeanAggregation extends AggregationFunction<GeometricMeanA
                 .compare(value, o.value)
                 .compare(n, o.n)
                 .result();
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) {
+                return true;
+            }
+            if (o == null || getClass() != o.getClass()) {
+                return false;
+            }
+            GeometricMeanState that = (GeometricMeanState) o;
+            return Objects.equals(value(), that.value());
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(value());
         }
     }
 
@@ -258,5 +289,128 @@ public class GeometricMeanAggregation extends AggregationFunction<GeometricMeanA
     @Override
     public Signature boundSignature() {
         return boundSignature;
+    }
+
+    @Override
+    public DocValueAggregator<?> getDocValueAggregator(List<DataType<?>> argumentTypes,
+                                                       List<MappedFieldType> fieldTypes) {
+        switch (argumentTypes.get(0).id()) {
+            case ByteType.ID:
+            case ShortType.ID:
+            case IntegerType.ID:
+            case LongType.ID:
+            case TimestampType.ID_WITH_TZ:
+            case TimestampType.ID_WITHOUT_TZ:
+                return new LongGeometricMean(fieldTypes.get(0).name());
+
+            case FloatType.ID:
+                return new FloatGeometricMean(fieldTypes.get(0).name());
+            case DoubleType.ID:
+                return new DoubleGeometricMean(fieldTypes.get(0).name());
+
+            default:
+                return null;
+        }
+    }
+
+    private static class LongGeometricMean implements DocValueAggregator<GeometricMeanState> {
+
+        private final String columnName;
+        private SortedNumericDocValues values;
+
+        public LongGeometricMean(String columnName) {
+            this.columnName = columnName;
+        }
+
+        @Override
+        public GeometricMeanState initialState() {
+            return new GeometricMeanState();
+        }
+
+        @Override
+        public void loadDocValues(LeafReader reader) throws IOException {
+            values = DocValues.getSortedNumeric(reader, columnName);
+        }
+
+        @Override
+        public void apply(GeometricMeanState state, int doc) throws IOException {
+            if (values.advanceExact(doc) && values.docValueCount() == 1) {
+                state.addValue(values.nextValue());
+            }
+        }
+
+        @Nullable
+        @Override
+        public Object partialResult(GeometricMeanState state) {
+            return state;
+        }
+    }
+
+    private static class DoubleGeometricMean implements DocValueAggregator<GeometricMeanState> {
+
+        private final String columnName;
+        private SortedNumericDocValues values;
+
+        public DoubleGeometricMean(String columnName) {
+            this.columnName = columnName;
+        }
+
+        @Override
+        public GeometricMeanState initialState() {
+            return new GeometricMeanState();
+        }
+
+        @Override
+        public void loadDocValues(LeafReader reader) throws IOException {
+            values = DocValues.getSortedNumeric(reader, columnName);
+        }
+
+        @Override
+        public void apply(GeometricMeanState state, int doc) throws IOException {
+            if (values.advanceExact(doc) && values.docValueCount() == 1) {
+                double value = NumericUtils.sortableLongToDouble(values.nextValue());
+                state.addValue(value);
+            }
+        }
+
+        @Nullable
+        @Override
+        public Object partialResult(GeometricMeanState state) {
+            return state;
+        }
+    }
+
+    private static class FloatGeometricMean implements DocValueAggregator<GeometricMeanState> {
+
+        private final String columnName;
+        private SortedNumericDocValues values;
+
+        public FloatGeometricMean(String columnName) {
+            this.columnName = columnName;
+        }
+
+        @Override
+        public GeometricMeanState initialState() {
+            return new GeometricMeanState();
+        }
+
+        @Override
+        public void loadDocValues(LeafReader reader) throws IOException {
+            values = DocValues.getSortedNumeric(reader, columnName);
+        }
+
+        @Override
+        public void apply(GeometricMeanState state, int doc) throws IOException {
+            if (values.advanceExact(doc) && values.docValueCount() == 1) {
+                double value = NumericUtils.sortableIntToFloat((int) values.nextValue());
+                state.addValue(value);
+            }
+        }
+
+        @Nullable
+        @Override
+        public Object partialResult(GeometricMeanState state) {
+            return state;
+        }
     }
 }