## Summary of the changes / Why this improves CrateDB

It improves the performance of the global standard deviation aggregations.
See https://github.com/crate/crate/pull/10048 for more information about the approach.

Without doc-values aggregators:

```
# Running setUp
# Running benchmark

## Running Query:
   Name: global geometric_mean on long
   Statement:
     select geometric_mean(duration) from uservisits
   Concurrency: 1
   Iterations: 5000
100%| 5000/5000 [04:53<00:00, 17.02 requests/s]
Runtime (in ms):
    mean:    56.871 ± 0.400
    min/max: 43.748 → 209.355
Percentile:
    50:   52.835 ± 14.417 (stdev)
    95:   79.098
    99.9: 195.374

## Running Query:
   Name: global geometric_mean on float
   Statement:
     select geometric_mean("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 5000
100%| 5000/5000 [05:14<00:00, 15.91 requests/s]
Runtime (in ms):
    mean:    61.897 ± 0.435
    min/max: 48.493 → 281.192
Percentile:
    50:   57.515 ± 15.699 (stdev)
    95:   86.598
    99.9: 194.373
# Running tearDown
```

With doc-values aggregators:

```
# Running setUp
# Running benchmark

## Running Query:
   Name: global geometric_mean on long
   Statement:
     select geometric_mean(duration) from uservisits
   Concurrency: 1
   Iterations: 5000
100%| 5000/5000 [02:59<00:00, 27.81 requests/s]
Runtime (in ms):
    mean:    34.720 ± 0.332
    min/max: 24.885 → 147.910
Percentile:
    50:   31.057 ± 11.964 (stdev)
    95:   54.400
    99.9: 128.699

## Running Query:
   Name: global geometric_mean on float
   Statement:
     select geometric_mean("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 5000
100%| 5000/5000 [02:45<00:00, 30.29 requests/s]
Runtime (in ms):
    mean:    31.969 ± 0.134
    min/max: 26.631 → 125.528
Percentile:
    50:   30.975 ± 4.850 (stdev)
    95:   39.082
    99.9: 58.210
# Running tearDown
```

## Checklist

 - [ ] Added an entry in `CHANGES.txt` for user facing changes
 - [ ] Updated documentation & `sql_features` table for user facing changes
 - [x] Touched code is covered by tests
 - [x] [CLA](https://crate.io/community/contribute/cla/) is signed
 - [x] This does not contain breaking changes, or if it does:
    - It is released within a major release
    - It is recorded in ``CHANGES.txt``
    - It was marked as deprecated in an earlier release if possible
    - You've thought about the consequences and other components are adapted
      (E.g. AdminUI)