@kovrus 2020-08-13
> Starting to wonder if we should have some SortedNumericDocValueAggregator base class that takes the columnName, and some Supplier<State> in the constructor. In most cases we'd just have to implement the apply function.

Yes, totally makes sense. I'll wrap up the rest of functions with the current approach, then will follow up on refactoring.